refactor: Add util.py (#3)
