+++ block_visualizer/src/block_visualizer/block_visualizer.py
@@ -1,44 +1,8 @@
-import json
-from datetime import datetime, date
 from pathlib import Path
-from typing import List, Dict, Any
-
 from jinja2 import Environment, FileSystemLoader
 
 from api.components import VisualizerPlugin
-
-
-class DateTimeEncoder(json.JSONEncoder):
-    """Custom JSON encoder that handles datetime and date objects."""
-
-    def default(self, obj: Any) -> str:
-        if isinstance(obj, (datetime, date)):
-            return obj.isoformat()
-        return super().default(obj)
-
-
-def _convert_nodes_to_dict(nodes) -> List[Dict[str, Any]]:
-    """Convert node objects to dictionary representation."""
-    return [
-        {
-            'id': node.get('id'),
-            **{k: v for k, v in node.items() if k != 'id'}
-        }
-        for node in nodes
-    ]
-
-
-def _convert_edges_to_dict(edges) -> List[Dict[str, str]]:
-    """Convert edge objects to dictionary representation, skipping invalid ones."""
-    return [
-        {"source": e["source"], "target": e["target"]}
-        for e in edges if "source" in e and "target" in e
-    ]
-
-
-def _custom_tojson(obj: Any, **kwargs) -> str:
-    """Custom JSON serialization filter for Jinja2 templates."""
-    return json.dumps(obj, cls=DateTimeEncoder, **kwargs)
+from .util import _convert_nodes_to_dict, _convert_edges_to_dict, _custom_tojson
 
 
 class BlockVisualizer(VisualizerPlugin):
